Move truncate tests to separate file
